Wave 48: 3880 -> 3930 (+50), bytes 57.12% -> 58.44%
Eleven agents, three concurrent, 50 matched of 66 attempted across
twelve address-locality blocks. Both builds reproduce
14dd0b22c894865867aff89e8116b2dffae25605.

WHAT AN AGENT INHERITS BEATS HOW BIG ITS BATCH IS. Eight batches,
all 3-6 functions under the standing cap, sorted by what the brief
gave them:

  briefed with a VERIFIED artifact       5/5, 5/5, 4/4, 8/8
  one block, dependency-ordered          9/10, 6/10, 3/4
  orchestrator's asm reading only        2/6

The verified artifacts were: a twin pair diffed by the orchestrator
before briefing (sub_08077140/sub_08077180 -- identical across all 37
instructions except `cmp #5` vs `cmp #3`); the byte-verified but
UNPROMOTED drafts of a block a previous agent had just finished,
named by work/<fn>/<fn>.c path because they are not in src/decomp/
yet; and a callee ordered ahead of its own caller so the prototype
was derived rather than guessed. The one 2/6 batch is the one where
the brief carried only the orchestrator's own reading of the
assembly. Size explained none of the spread.

proto_check.py's DUPLICATE-TAG CHECK WAS BLIND TO 115 OF 180 TAG
DECLARATIONS (64%) AND REPORTED CLEAN THROUGHOUT A REPO-WIDE BREAK.
Its regex required end-of-line after the tag, while this project's
own convention writes `struct Unk0200C078Rec /* 0x04 */`. All three
tags that were actually duplicated carry a size comment, so none was
visible. Fixed and verified in both directions. A checker that cannot
see two thirds of its own corpus reads exactly like an all-clear --
sixth wave disrupted by tooling rather than decompilation. Cheap
independent cross-check, one second, no tool required:
  grep -nE '^struct Unk[0-9A-Za-z]+ ' include/unknown-globals.h \
    | awk '{print $2}' | sort | uniq -d

The break itself: a TWO-ANCHOR SPLICE whose `end` anchor matched an
earlier identical line, replaying 269 lines and giving three struct
tags two divergent bodies each. Every try_match in the repo failed
until repaired. The competing "whole-file rewrite" theory was wrong
-- the splice mechanism predicts the identical 7-line and 84-line tag
spacing measured in both copies, and a rewrite does not. Two distinct
lessons, both worth keeping: edit shared headers IN PLACE, and check
both anchors are unique before splicing. No work was lost; the one
function reported blocked by it was correct all along.

A -fforce-addr POOL WORD IS NOT A GLOBAL, CONFIRMED TWICE BY
DEREFERENCING baserom.gba RATHER THAN TRUSTING THE SYMBOL NAME. The
ROM word at 0x0816E16C holds 0x0202F214 and 0x0816D93C holds
0x08499590 -- gen_lds.py invented symbols for both. sub_080585D4
reads the map header's width through gUnknown_08499590 and its height
through the -fforce-addr spelling of the SAME object. Declaring one
of these creates a second name for an object that already has one,
and a wrong-but-byte-neutral type has no oracle at all: no build, no
try_match, no ROM hash catches it. The first instance identified a
producer/consumer pair that no other screen could have found.

Codegen rules added to docs/agbcc-codegen.md, each from a measurement:
- Bitfields: agbcc gives ANY struct containing a bitfield 4-byte SIZE
  AND ALIGNMENT whatever the field's base type, with a five-layout
  table. So a bitfield often cannot sit where the ROM's stride says
  the data is, which rules the reading out without spending a
  try_match. Corollary: `ldrb` + `lsls #0x1e / lsrs #0x1e` is an
  explicit (u8) narrowing, NOT a bitfield and not `& 3`.
- check_dbra_loop: the preheader-POSITION rule read BACKWARDS is a
  loop-DIRECTION discriminator, now extended to outermost loops --
  measure the counter's init against the LICM hoists and the giv init
  in its own preheader. sub_0806A6F0 and sub_0806A5B8 are an opposing
  pair in ONE block: both ROMs count down, both end `cmp rN,#0; bge`,
  so nothing at the loop bottom separates them, and the source loops
  are ascending and descending respectively.
- p[X + C], loop form: `p + i*S + C` gives a giv of `i*S` and re-adds
  C at every use; `p + (i*S + C)` folds C into the giv's init. AN
  UNEXPLAINED `sub sp, #N` AND A MISPLACED CONSTANT ADD ARE ONE FACT
  -- the two-giv form ran sub_0806BF40 out of callee-saved registers
  and produced the ROM's spill. Grouping alone took it to byte-exact.

sub_0806B120 parked, and NOT for lack of a source model. It is the
WRITER of gUnknown_0202F214 and sub_0806AD04 -- matched this wave --
is the READER, and they want incompatible declarations of the field
at +0x02: the writer's stores are textbook store_bit_field and need
bitfields, the reader's loads cannot use them because bitfield reads
widen to the enclosing word, and a union view inflates the element
from 4 to 8 bytes and breaks the stride. Declining to reshape a
struct an already-matched function reads is the rule that has broken
the build in five separate waves when ignored. Next wave must budget
it as ONE coupled job over both functions with sub_0806AD04
re-verified, not as a 136-byte single.

Also: carved gUnknown_084C37E4 out of the covering .incbin in
data-0848B688.s (0x11C -> 0x100 + 0x1C) -- an undefined ROM reference
fails ONLY the split link while the assembly build prints
`aw2bhr.gba: OK` beside the failure. Removed a STALE parked entry for
sub_08070BAC, matched in wave 47; a stale entry once suppressed
AgbMain from three waves of pre-flights.

A batching note for the next wave: only 34 of the block screen's 137
candidates have no work/ directory. Several blocks that look open are
full of 93-97% drafts from earlier waves, which are measured to
convert at roughly zero. Screen on draft ABSENCE before adjacency.

+++ src/decomp/c_0805ACA8.c
@@ -59,7 +59,7 @@ struct Map
  *
  * The fourth call's `adds r4,#1` clobbers y because y is dead after it; that
  * falls out of writing `y + 1` last. */
-int sub_0805ACA8(int x, int y, u16 * out)
+u8 sub_0805ACA8(int x, int y, u16 * out)
 {
     int sentinel;
